refactor: moved LottoNumberGenerator to domain package
